refactor: convert worker to typescript module syntax
