+++ .gitignore
@@ -9,3 +9,4 @@ worker/
 node_modules/
 .cargo-ok
 .wrangler/
+worker-configuration.d.ts
