+++ .prettierignore
@@ -1,0 +1,2 @@
+worker-configuration.d.ts
+package-lock.json
