--- src/util.js
@@ -1,7 +1,0 @@
-export function randomIndex(array) {
-    return Math.floor(Math.random() * array.length);
-}
-
-export function randomEntry(array) {
-    return array[randomIndex(array)];
-}
